# stronger/server/docs/swagger.yaml
basePath: /api/
definitions:
...
  server_module_pro_manager_model.Announcement:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      proCode:
        type: string
      releaseDate:
        type: string
      releaseType:
        type: integer
      releaseUserCode:
        type: string
      releaseUserName:
        type: string
      status:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
      visitCount:
        type: integer
    type: object
...
  server_module_pro_manager_model_request.AnnouncementChangeStatusReq:
    properties:
      ids:
        items:
          type: string
        type: array
      status:
        description: 删除传：3，发布传：2，取消发布和恢复传：1
        type: integer
    required:
    - status
    type: object
...
info:
  contact: {}
  description: This is stronger Server pets
  title: stronger doc
  version: 1.0.1
paths:
...
  /homepage/home/announcement:
    get:
      consumes:
      - application/json
      parameters:
      - description: 开始时间格式2022-07-06T16:00:00.000Z
        in: query
        name: startTime
        type: string
      - description: 结束时间格式2022-07-06T16:00:00.000Z
        in: query
        name: endTime
        type: string
      - description: '发布类型1: 公告通知, 2: 规则动态'
        in: query
        name: releaseType
        required: true
        type: integer
      - description: 标题
        in: query
        name: title
        type: string
      - description: 项目
        in: query
        name: proCode
        type: string
      - description: 页码
        in: query
        name: pageIndex
        required: true
        type: integer
      - description: 数量
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 首页--主页获取公告
      tags:
      - homepage(首页--我的主页)
  /homepage/home/announcement-view:
    get:
      consumes:
      - application/json
      parameters:
      - description: 公告id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 首页--新增访问量
      tags:
      - homepage(首页--我的主页)
...
  /pro-manager/announcement-manager/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 公告实体类
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/server_module_pro_manager_model.Announcement'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 项目管理--新增一个公告
      tags:
      - pro-manager(项目管理--公告管理)
  /pro-manager/announcement-manager/change-status:
    post:
      consumes:
      - application/json
      parameters:
      - description: id数组
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/server_module_pro_manager_model_request.AnnouncementChangeStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 项目管理--删除、恢复(待发布)、发布、取消发布公告
      tags:
      - pro-manager(项目管理--公告管理)
  /pro-manager/announcement-manager/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 公告实体类
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/server_module_pro_manager_model.Announcement'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 项目管理--编辑公告
      tags:
      - pro-manager(项目管理--公告管理)
  /pro-manager/announcement-manager/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 开始时间格式2022-07-06T16:00:00.000Z
        in: query
        name: startTime
        type: string
      - description: 结束时间格式2022-07-06T16:00:00.000Z
        in: query
        name: endTime
        type: string
      - description: 项目编码
        in: query
        name: proCode
        type: string
      - description: 发布类型
        in: query
        name: releaseType
        type: integer
      - description: 发布状态
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: pageIndex
        required: true
        type: integer
      - description: 数量
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 排序JSON.stringify([[
        in: query
        name: orderBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 项目管理--获取公告管理分页
      tags:
      - pro-manager(项目管理--公告管理)
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
  ProCode:
    in: header
    name: pro-code
    type: apiKey
  UserID:
    in: header
    name: x-user-id
    type: apiKey
swagger: "2.0"